chore(discussion/client): rename client files
