+++ pkg/cmd/discussion/client/client_test.go
